Added minor edits to a_star function ("a_star.cpp" and "algorithms.h files).
Edited it because/so that it has the same function parameters as the rest of the maze solving algorithms in order to further refactor/clean up the "main.cpp"'s code later.
This single commit is only for diving 1 bigger commit into 2 smaller ones (so it's experiemental and probably will not work/pass tests).
(Clean the "a_star.cpp" code more and fix the not proper full solution node path storing later!!!)